DjangoBackend/requirements.txt
absl-py==1.1.0
aiohttp==3.8.2
aiosignal==1.3.1
annotated-types==0.6.0
appnope==0.1.3
APScheduler==3.10.2
asgiref==3.7.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==1.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==41.0.5
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.8.0rc2
deprecation==2.1.0
distlib==0.3.4
Django==4.2.7
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
dynaconf==3.2.4
entrypoints==0.4
eventkit==1.0.0
executing==0.8.3
filelock==3.7.1
flatbuffers==1.12
frozenlist==1.4.0
gast==0.4.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.7.0
hvac==2.0.0
ib-insync==0.9.86
idna==3.3
install==1.3.5
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
knox==0.1.14
libclang==14.0.1
loguru==0.7.2
Markdown==3.3.7
matplotlib-inline==0.1.3
msgpack==1.0.3
multidict==5.2.0
nest-asyncio==1.5.5
networkx==3.1
nltk==3.7
numpy==1.22.4
oauthlib==3.2.0
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.12.0
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
pyzmq==23.1.0
redis==4.6.0
regex==2022.6.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
shortuuid==1.0.11
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
spotify==0.10.2
spotipy==2.23.0
sqlparse==0.4.4
sseclient-py==1.8.0
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tflearn==0.5.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.9
validators==0.22.0
virtualenv==20.14.1
wcwidth==0.2.5
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==2.1.2
wrapt==1.14.1
xlrd==2.0.1
yarl==1.9.2
